+++ ide-extensions~/lsp-server/obj/Debug/net8.0/UitkxLanguageServer.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("UitkxLanguageServer")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+7da6720045ac6c3f16d9e8ae2fa0eb25ba1b4c81")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+17f2a9ce1d46c535740d0d209b6586f8d98ff552")]
 [assembly: System.Reflection.AssemblyProductAttribute("UitkxLanguageServer")]
 [assembly: System.Reflection.AssemblyTitleAttribute("UitkxLanguageServer")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]